修复jsx中使用getFieldDecorator报错Cannot read property 'value' of undefined
